# BikeWorldServer/api-docs.yaml
openapi: 3.0.0
info:
  title: bike-world API
  version: '2.0'
servers:
  - url: https://bike-world-ids.herokuapp.com/
paths:
...
  /api/v1/rentals/{id}:
    put:
      summary: Update a rental point.
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the rental point to update.
          schema:
            type: string
            example: '627fb874a991e0a164e063ac'
        - name: x-access-token
          in: header
          required: true
          description: The jwt of the session.
          schema:
            type: string
            example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'Est Milano'
                address:
                  type: string
                  example: 'ds'
                lat:
                  type: number
                  example: 44.9654219
                lng:
                  type: number
                  example:  9.1859243
                type:
                  type: string
                  example: "Negozio"
      responses:
        '200':
          description: Rental point updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Rental point info updated!"
        '404':
          description: Rental not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Rental Point not found"
        '401':
          description: Unauthorized. No token provided.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'No token provided.'
        '403':
          description: Forbidden. Failed to authenticate token.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Failed to authenticate token.'
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'
  /api/v1/rentals/type:
    get:
      summary: Get rental point by type.
      parameters:
        - name: type
          in: query
          required: true
          description: The type of the rental points to return.
          schema:
            type: string
            example: 'Negozio'
      responses:
        '200':
          description: List of the rental points with a specific type.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id: 
                      type: string
                      example: "627fb874a991e0a164e063ac"
                    name:
                      type: string
                      example: "Est Milano"
                    address:
                      type: string
                      example: "ds"
                    lat: 
                      type: number
                      example: 45.4654219
                    lng: 
                      type: number
                      example: 10.1859243
                    type:
                      type: string
                      example: "Negozio"
                    bikeNumber:
                      type: integer
                      example: 0
                    self:
                      type: string
                      example: "/api/v1/rentals/627fb874a991e0a164e063ac"  
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'
  /api/v1/rentals/zone:
    get:
      summary: Get rental point by zone.
      parameters:
        - name: latitude
          in: query
          required: true
          description: The latitude of the rental points to return.
          schema:
            type: number
            example: 45.4654219
        - name: longitude
          in: query
          required: true
          description: The longitude of the rental points to return.
          schema:
            type: number
            example: 10.1859243
      responses:
        '200':
          description: List of the rental points with a specific type.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id: 
                      type: string
                      example: "627fb874a991e0a164e063ac"
                    name:
                      type: string
                      example: "Est Milano"
                    address:
                      type: string
                      example: "ds"
                    lat: 
                      type: number
                      example: 45.4654219
                    lng: 
                      type: number
                      example: 10.1859243
                    type:
                      type: string
                      example: "Negozio"
                    bikeNumber:
                      type: integer
                      example: 0
                    self:
                      type: string
                      example: "/api/v1/rentals/627fb874a991e0a164e063ac"  
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'
  /api/v1/rentals/date:
    get:
      summary: Get rental point based on the bike availability.
      parameters:
        - name: date
          in: query
          required: true
          description: The date used to check the bike availability.
          schema:
            type: string
            example: '2022-05-20'
      responses:
        '200':
          description: List of the rental points that are available in a specific date.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id: 
                      type: string
                      example: "627fb874a991e0a164e063ac"
                    name:
                      type: string
                      example: "Est Milano"
                    address:
                      type: string
                      example: "ds"
                    lat: 
                      type: number
                      example: 45.4654219
                    lng: 
                      type: number
                      example: 10.1859243
                    type:
                      type: string
                      example: "Negozio"
                    bikeNumber:
                      type: integer
                      example: 0
                    self:
                      type: string
                      example: "/api/v1/rentals/627fb874a991e0a164e063ac"  
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'

  ###################
  # route bookings
  ###################
...
  /api/v2/rentals/{id}:
    put:
      summary: Update a rental point.
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the rental point to update.
          schema:
            type: string
            example: '627fb874a991e0a164e063ac'
        - name: x-access-token
          in: header
          required: true
          description: The jwt of the session.
          schema:
            type: string
            example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'Est Milano'
                address:
                  type: string
                  example: 'ds'
                lat:
                  type: number
                  example: 44.9654219
                lng:
                  type: number
                  example:  9.1859243
                type:
                  type: string
                  example: "Negozio"
      responses:
        '200':
          description: Rental point updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Rental point info updated!"
        '404':
          description: Rental not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Rental Point not found"
        '401':
          description: Unauthorized. No token provided.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'No token provided.'
        '403':
          description: Forbidden. Failed to authenticate token.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Failed to authenticate token.'
        '400':
          description: Bad Request. Check docs for required parameters. /api/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/api-docs'
  /api/v2/rentals/type:
    get:
      summary: Get rental point by type.
      parameters:
        - name: type
          in: query
          required: true
          description: The type of the rental points to return.
          schema:
            type: string
            example: 'Negozio'
      responses:
        '200':
          description: List of the rental points with a specific type.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id: 
                      type: string
                      example: "627fb874a991e0a164e063ac"
                    name:
                      type: string
                      example: "Est Milano"
                    address:
                      type: string
                      example: "ds"
                    lat: 
                      type: number
                      example: 45.4654219
                    lng: 
                      type: number
                      example: 10.1859243
                    type:
                      type: string
                      example: "Negozio"
                    bikeNumber:
                      type: integer
                      example: 0
                    self:
                      type: string
                      example: "/api/v2/rentals/627fb874a991e0a164e063ac"  
        '400':
          description: Bad Request. Check docs for required parameters. /api/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/api-docs'
  /api/v2/rentals/zone:
    get:
      summary: Get rental point by zone.
      parameters:
        - name: latitude
          in: query
          required: true
          description: The latitude of the rental points to return.
          schema:
            type: number
            example: 45.4654219
        - name: longitude
          in: query
          required: true
          description: The longitude of the rental points to return.
          schema:
            type: number
            example: 10.1859243
      responses:
        '200':
          description: List of the rental points with a specific type.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id: 
                      type: string
                      example: "627fb874a991e0a164e063ac"
                    name:
                      type: string
                      example: "Est Milano"
                    address:
                      type: string
                      example: "ds"
                    lat: 
                      type: number
                      example: 45.4654219
                    lng: 
                      type: number
                      example: 10.1859243
                    type:
                      type: string
                      example: "Negozio"
                    bikeNumber:
                      type: integer
                      example: 0
                    self:
                      type: string
                      example: "/api/v2/rentals/627fb874a991e0a164e063ac"  
        '400':
          description: Bad Request. Check docs for required parameters. /api/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/api-docs'
  /api/v2/rentals/date:
    get:
      summary: Get rental point based on the bike availability.
      parameters:
        - name: date
          in: query
          required: true
          description: The date used to check the bike availability.
          schema:
            type: string
            example: '2022-05-20'
      responses:
        '200':
          description: List of the rental points that are available in a specific date.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id: 
                      type: string
                      example: "627fb874a991e0a164e063ac"
                    name:
                      type: string
                      example: "Est Milano"
                    address:
                      type: string
                      example: "ds"
                    lat: 
                      type: number
                      example: 45.4654219
                    lng: 
                      type: number
                      example: 10.1859243
                    type:
                      type: string
                      example: "Negozio"
                    bikeNumber:
                      type: integer
                      example: 0
                    self:
                      type: string
                      example: "/api/v2/rentals/627fb874a991e0a164e063ac"  
        '400':
          description: Bad Request. Check docs for required parameters. /api/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/api-docs'

  ###################
  # route bookings
  ###################
...